fix: some build problems, utils usage problems etc
