+++ apps/server/package.json
@@ -14,9 +14,10 @@
   },
   "dependencies": {
     "@chatty/types": "*",
+    "@types/bcryptjs": "^2.4.2",
     "@types/jest": "^29.5.0",
     "@types/mocha": "^10.0.1",
-    "bcrypt": "^5.1.0",
+    "bcryptjs": "^2.4.3",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
