+++ apps/server/src/utils/bcrypt/index.ts
@@ -1,10 +1,10 @@
-import bcrypt from 'bcrypt';
+import { hash, compare } from 'bcryptjs';
 
 export const hashPassword = async (password: string) => {
-  const hashed = await bcrypt.hash(password, 12);
+  const hashed = await hash(password, 12);
 
   return hashed;
 };
 
 export const comparePasswords = (password: string, hashed: string) =>
-  bcrypt.compare(password, hashed);
+  compare(password, hashed);
