+++ packages/types/enums/statusCodes.enum.ts
@@ -1,0 +1,7 @@
+export const STATUS_CODES = {
+  CONFLICT: 409,
+  CREATED: 201,
+  SUCCESS: 200,
+  WRONG_DATA: 400,
+  NOT_FOUND: 404,
+};
